Add Activityindicator while Whois-Request
Close #1 

NOTE: The fetch-Process was pretty fast at my localhost. So I had no chance to see the loading-indicator.
But to give it a try, you can add an setTimeout to line 32 and give it a try by typing in an .com-domain. Then you see the indicator for a second.
```
setTimeout(()=>this.toggleActivityIndicator('hide'),1000);
```
So if the user has an delay while fetching data, he will see the indicator, if not he wouldn't recognize that there is one (I think that's the best solution for UX).